fix/refactor: first review - разбить Mapper по соответствующим сущностям
