+++ src/main/java/ru/practicum/shareit/item/mapper/CommentMapper.java
@@ -1,0 +1,33 @@
+package ru.practicum.shareit.item.mapper;
+
+import org.springframework.stereotype.Component;
+import ru.practicum.shareit.item.dto.CommentDto;
+import ru.practicum.shareit.item.dto.CommentTextDto;
+import ru.practicum.shareit.item.model.Comment;
+import ru.practicum.shareit.item.model.Item;
+import ru.practicum.shareit.user.model.User;
+
+import java.time.LocalDateTime;
+
+@Component
+public class CommentMapper {
+
+    public Comment toComment(CommentTextDto comment, Item item, User author) {
+        return new Comment(
+                0,
+                comment.getText(),
+                item,
+                author,
+                LocalDateTime.now()
+        );
+    }
+
+    public CommentDto toDto(Comment comment) {
+        return new CommentDto(
+                comment.getId(),
+                comment.getText(),
+                comment.getAuthor().getName(),
+                comment.getCreated()
+        );
+    }
+}
