+++ src/main/java/ru/practicum/shareit/user/mapper/UserMapper.java
@@ -1,0 +1,25 @@
+package ru.practicum.shareit.user.mapper;
+
+import org.springframework.stereotype.Component;
+import ru.practicum.shareit.user.dto.UserDto;
+import ru.practicum.shareit.user.model.User;
+
+@Component
+public class UserMapper {
+
+    public UserDto toDto(User user) {
+        return new UserDto(
+                user.getId(),
+                user.getName(),
+                user.getEmail()
+        );
+    }
+
+    public User toUser(UserDto userDto) {
+        return new User(
+                userDto.getId(),
+                userDto.getName(),
+                userDto.getEmail()
+        );
+    }
+}
